+++ internal/api/helpers.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"net/url"
 
 	"github.com/gofrs/uuid"
 	"github.com/pkg/errors"
@@ -61,73 +60,6 @@ func (a *API) requestAud(ctx context.Context, r *http.Request) string {
 
 	// Finally, return the default if none of the above methods are successful
 	return config.JWT.Aud
-}
-
-// tries extract redirect url from header or from query params
-func getRedirectTo(r *http.Request) (reqref string) {
-	reqref = r.Header.Get("redirect_to")
-	if reqref != "" {
-		return
-	}
-
-	if err := r.ParseForm(); err == nil {
-		reqref = r.Form.Get("redirect_to")
-	}
-
-	return
-}
-
-func isRedirectURLValid(config *conf.GlobalConfiguration, redirectURL string) bool {
-	if redirectURL == "" {
-		return false
-	}
-
-	base, berr := url.Parse(config.SiteURL)
-	refurl, rerr := url.Parse(redirectURL)
-
-	// As long as the referrer came from the site, we will redirect back there
-	if berr == nil && rerr == nil && base.Hostname() == refurl.Hostname() {
-		return true
-	}
-
-	// For case when user came from mobile app or other permitted resource - redirect back
-	for _, pattern := range config.URIAllowListMap {
-		if pattern.Match(redirectURL) {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (a *API) getReferrer(r *http.Request) string {
-	config := a.config
-
-	// try get redirect url from query or post data first
-	reqref := getRedirectTo(r)
-	if isRedirectURLValid(config, reqref) {
-		return reqref
-	}
-
-	// instead try referrer header value
-	reqref = r.Referer()
-	if isRedirectURLValid(config, reqref) {
-		return reqref
-	}
-
-	return config.SiteURL
-}
-
-// getRedirectURLOrReferrer ensures any redirect URL is from a safe origin
-func (a *API) getRedirectURLOrReferrer(r *http.Request, reqref string) string {
-	config := a.config
-
-	// if redirect url fails - try fill by extra variant
-	if isRedirectURLValid(config, reqref) {
-		return reqref
-	}
-
-	return a.getReferrer(r)
 }
 
 func isStringInSlice(checkValue string, list []string) bool {
