+++ internal/utilities/request.go
@@ -5,7 +5,10 @@ import (
 	"io"
 	"net"
 	"net/http"
+	"net/url"
 	"strings"
+
+	"github.com/supabase/gotrue/internal/conf"
 )
 
 // GetIPAddress returns the real IP address of the HTTP request. It parses the
@@ -53,4 +56,57 @@ func GetBodyBytes(req *http.Request) ([]byte, error) {
 	req.Body = io.NopCloser(bytes.NewReader(buf))
 
 	return buf, nil
+}
+
+func GetReferrer(r *http.Request, config *conf.GlobalConfiguration) string {
+	// try get redirect url from query or post data first
+	reqref := getRedirectTo(r)
+	if IsRedirectURLValid(config, reqref) {
+		return reqref
+	}
+
+	// instead try referrer header value
+	reqref = r.Referer()
+	if IsRedirectURLValid(config, reqref) {
+		return reqref
+	}
+
+	return config.SiteURL
+}
+
+func IsRedirectURLValid(config *conf.GlobalConfiguration, redirectURL string) bool {
+	if redirectURL == "" {
+		return false
+	}
+
+	base, berr := url.Parse(config.SiteURL)
+	refurl, rerr := url.Parse(redirectURL)
+
+	// As long as the referrer came from the site, we will redirect back there
+	if berr == nil && rerr == nil && base.Hostname() == refurl.Hostname() {
+		return true
+	}
+
+	// For case when user came from mobile app or other permitted resource - redirect back
+	for _, pattern := range config.URIAllowListMap {
+		if pattern.Match(redirectURL) {
+			return true
+		}
+	}
+
+	return false
+}
+
+// getRedirectTo tries extract redirect url from header or from query params
+func getRedirectTo(r *http.Request) (reqref string) {
+	reqref = r.Header.Get("redirect_to")
+	if reqref != "" {
+		return
+	}
+
+	if err := r.ParseForm(); err == nil {
+		reqref = r.Form.Get("redirect_to")
+	}
+
+	return
 }
